feat: optional system runtime statistics (#28)
Realtime runtime statistics can be enabled with IDF build flags.
This is a development feature to monitor CPU & stack usage.

Once enabled, statistics are continuously gathered in an interval of 2 seconds.
The CPU & free stack usage of all tasks are calculated.
The stats are logged in the console and can also be retrieved with a WebSocket request `get_stats`.