feat: sending an IR code for a specific time
Some devices require pressing a button for a specific time to trigger a different action.
For example long pressing the power button will switch off the device.

This could be simulated with a proper repeat count. But this requires to know the IR protocol specific timings

The new optional `hold` field in the `ir_send` message allows to send the code for a specific time.
Important: this time is the **minimal duration** the IR signal is sent and not the exact time the signal is stopped.
An IR message will always be completed. This means that the actual send time will mostly be longer.

- The `hold` parameter will override the `repeat` field if present!
  - Total signal time: MAX(hold, "minimal IR signal duration")
  - Example: Sony (protocol 4) requires a minimum of 3 IR messages, each 45ms apart:
    - minimal signal duration = 135ms
    - any `hold` value < 135 is ignored and the IR signal will be active for 135ms
    - if `hold` is set to 150: the signal will be active for 180ms because 4 x 45ms = 180ms
- If using a `format: hex` message:
  - The repeat count in `code` is ignored.
  - The protocol specific minimal repeat count is still being used. For example Sony (protocol 4), uses a minimal repeat count of 2.