fix: LVGL task locks during startup with led indicator running
The UI state machine is driven by the default timer task which has a task
priority of 1. LVGL task uses a priority of 4.
During startup the display clear call at the end of the info screens frequently
triggered an LVGL lock error.
Even increasing the lock timeout to 20s didn't solve it.

By setting the timer task to the same priority 4, these issues seem to be gone.
It could no longer be reproduced by restarting the dock > 60 times.
Without that change the lock appeared every 3-5 startups while a led_indicator
task was running.